Revert "Milestone 2 - Changing medium directives to components and test all to get 100% coverage (Part 2) (#9961)"
This reverts commit 5e39aba6d86a6d03fe1fb67e2384184f58429bc7.
