+++ core/templates/pages/exploration-editor-page/editor-navigation/editor-navigation.directive.html
