+++ core/templates/pages/exploration-editor-page/editor-navigation/editor-navigation.directive.ts
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 /**
- * @fileoverview Component for showing Editor Navigation
+ * @fileoverview Directive for showing Editor Navigation
  * in editor.
  */
 
@@ -32,8 +32,9 @@ require('services/site-analytics.service.ts');
 require('services/user.service.ts');
 require('services/contextual/window-dimensions.service.ts');
 
-angular.module('oppia').component('editorNavigation', {
-  template: require('./editor-navigation.component.html'),
+angular.module('oppia').directive('editorNavigation', () => ({
+  restrict: 'E',
+  template: require('./editor-navigation.directive.html'),
   controller: [
     '$q', '$rootScope', '$scope', '$timeout', '$uibModal', 'ContextService',
     'ExplorationImprovementsService', 'ExplorationRightsService',
@@ -135,5 +136,5 @@ angular.module('oppia').component('editorNavigation', {
         }
       };
     }
-  ]
-});
+  ],
+}));
