+++ core/templates/pages/exploration-editor-page/exploration-editor-page.scripts.ts
@@ -28,7 +28,7 @@ require(
   'editor-navbar-breadcrumb.directive.ts');
 require(
   'pages/exploration-editor-page/editor-navigation/' +
-  'editor-navigation.component.ts');
+  'editor-navigation.directive.ts');
 require(
   'pages/exploration-editor-page/exploration-save-and-publish-buttons/' +
   'exploration-save-and-publish-buttons.directive.ts');
