+++ core/templates/pages/exploration-editor-page/translation-tab/state-translation-editor/state-translation-editor.directive.html
