+++ scripts/check_frontend_coverage.py
@@ -110,6 +110,7 @@ NOT_FULLY_COVERED_FILENAMES = [
     'editable-collection-backend-api.service.ts',
     'editable-story-backend-api.service.ts',
     'editor-navbar-breadcrumb.directive.ts',
+    'editor-navigation.directive.ts',
     'email-dashboard-data.service.ts',
     'entity-creation.service.ts',
     'exploration-automatic-text-to-speech.service.ts',
@@ -273,6 +274,7 @@ NOT_FULLY_COVERED_FILENAMES = [
     'Polyfills.ts',
     'practice-tab.directive.ts',
     'pretest-question-backend-api.service.ts',
+    'preview-tab.directive.ts',
     'profile-link-image.directive.ts',
     'profile-link-text.directive.ts',
     'profile-page-navbar.component.ts',
@@ -362,6 +364,7 @@ NOT_FULLY_COVERED_FILENAMES = [
     'state-responses.directive.ts',
     'state-solution-editor.directive.ts',
     'state-top-answers-stats.service.ts',
+    'state-translation-editor.directive.ts',
     'state-translation-status-graph.directive.ts',
     'state-translation.directive.ts',
     'state-tutorial-first-time.service.ts',
@@ -423,6 +426,7 @@ NOT_FULLY_COVERED_FILENAMES = [
     'truncate.pipe.ts',
     'tutor-card.directive.ts',
     'unicode-string-editor.directive.ts',
+    'unresolved-answers-overview.directive.ts',
     'url-interpolation.service.ts',
     'utils.service.ts',
     'version-diff-visualization.directive.ts',
